optional auth set in web's Dockerfile
