+++ web/src/auth/AuthContext.jsx
@@ -3,8 +3,17 @@ import React from 'react';
 export const AuthContext = React.createContext();
 
 export const AuthProvider = ({idToken, children}) => {
+  let initialState = idToken
+  if(!initialState){
+    initialState = {
+      user: "Not Logged in",
+      profile: {
+        name: "Not Authenticated"
+      }
+    }
+  }
   return (
-    <AuthContext.Provider value={idToken}>{children}</AuthContext.Provider>
+    <AuthContext.Provider value={initialState}>{children}</AuthContext.Provider>
   );
 };
