+++ web/src/auth/adalConfig.js
@@ -1,10 +1,10 @@
 import { AuthenticationContext } from 'react-adal'
 
 export const adalConfig = {
-  tenant: '3aa4a235-b6e2-48d5-9195-7fcf05b459b0',
-  clientId: '97a6b5bd-63fb-42c6-bb75-7e5de2394ba0',
+  tenant: process.env.REACT_APP_AUTH_TENANT,
+  clientId: process.env.REACT_APP_AUTH_CLIENT_ID,
   endpoints: {
-    api: '97a6b5bd-63fb-42c6-bb75-7e5de2394ba0',
+    api: process.env.REACT_APP_AUTH_CLIENT_ID,
   },
   cacheLocation: 'localStorage',
 }
